refactor: split model.go into per-concern files and rename processor_impl.go
- internal/model/model.go removed; split into:
  - article.go  (Article, ProcessedArticle, LocaleRef, FrontMatter)
  - config.go   (Config, SiteConfig, BuildConfig, ThemeConfig,
                  SyntaxHighlightConfig, OGPConfig, I18nConfig)
  - site.go     (Site, Pagination, FileWatcher, DevServer)
  - manifest.go (BuildManifest, OutputFile)
  - taxonomy.go (Taxonomy, TaxonomyRegistry)
  - graph.go    (NodeType, Node, DependencyGraph, ChangeSet)
- internal/processor/processor_impl.go renamed to site.go;
  _impl suffix is not idiomatic Go
