+++ internal/model/article.go
@@ -1,0 +1,64 @@
+package model
+
+import (
+	"html/template"
+	"time"
+)
+
+// Article is the raw data parsed from a Markdown file.
+type Article struct {
+	FrontMatter  FrontMatter
+	RawContent   string
+	FilePath     string
+	LastModified time.Time
+}
+
+// ProcessedArticle holds derived data generated by the renderer.
+type ProcessedArticle struct {
+	Article
+	HTMLContent template.HTML
+	Summary     string
+	OutputPath  string
+	// ContentPath is the content-dir-relative path to the source Markdown file
+	// (e.g. "posts/hello-world.md"). Used to generate GitHub edit/view links.
+	ContentPath string
+	// Locale is the locale code detected from the content path (e.g. "en", "ja").
+	// Empty when i18n is not configured.
+	Locale string
+	// URL is the canonical URL path for this article (e.g. "/posts/hello/" or
+	// "/ja/posts/hello/"). Empty when i18n is not configured.
+	URL string
+	// Translations lists translated variants of this article, keyed by locale.
+	// Populated by BuildTranslationMap after article processing.
+	Translations []LocaleRef
+	// PluginData holds per-article data injected by enabled plugins.
+	// Access in templates: {{index .PluginData "amazon_books"}}
+	PluginData map[string]interface{}
+}
+
+// LocaleRef holds a locale code and the canonical URL for a translated variant
+// of an article. Used to generate language-switcher links in templates.
+type LocaleRef struct {
+	Locale string
+	URL    string
+}
+
+// FrontMatter holds the YAML metadata from the top of a Markdown file.
+type FrontMatter struct {
+	Title       string    `yaml:"title"`
+	Date        time.Time `yaml:"date"`
+	Draft       bool      `yaml:"draft"`
+	Tags        []string  `yaml:"tags"`
+	Categories  []string  `yaml:"categories"`
+	Description string    `yaml:"description"`
+	Author      string    `yaml:"author"`
+	Slug        string    `yaml:"slug"`
+	Template    string    `yaml:"template"`
+	// TranslationKey links this article to its translations in other locales.
+	// Articles sharing the same key are treated as translations of each other,
+	// enabling language-switcher links via ProcessedArticle.Translations.
+	TranslationKey string `yaml:"translation_key"`
+	// Extra captures any front-matter keys not listed above.
+	// Plugins read their configuration from this field.
+	Extra map[string]interface{} `yaml:",inline"`
+}
